fix: Removal of step 4 in the registration flow and validation of data saved in the registration localstorage
