+++ producer/src/app/cadastrar/data/index.ts
@@ -10,5 +10,4 @@ export const steps = [
   2,
   3,
   4,
-  5
 ]
