+++ producer/src/app/layout.tsx
@@ -4,6 +4,7 @@ import { Toaster } from "sonner";
 
 import { setAppID } from "@shared/next/library/set-app-id";
 import { getAppID } from "@shared/next/library/get-app-id";
+import ClientValidation from "@shared/components/ClientValidation"
 
 export const metadata: Metadata = {
   title: "Painel e-COO",
@@ -22,7 +23,11 @@ export default function RootLayout({
       <body className={`${getAppID()} w-screen h-screen font-poppins`}>
         <div className="flex flex-row justify-center w-full h-[inherit]">
           <Toaster richColors position="top-right" />
-          <div className="relative max-w-md w-full h-[inherit]">{children}</div>
+          <div className="relative max-w-md w-full h-[inherit]">
+            <ClientValidation>
+              {children}
+            </ClientValidation>
+          </div>
         </div>
       </body>
     </html>
