+++ shared/src/errors.ts
@@ -1,4 +1,4 @@
-  export const errorsMapper: { [key: string]: string } = {
+  export const errorsMapper: Record<string, string> = {
     "Sessão expirada.": "Sua sessão expirou. Por favor, faça login novamente.",
     "💥 Ocorreu um erro interno.": "Ocorreu um erro interno. Tente novamente mais tarde.",
     "As credenciais de acesso não são válidas.": "Suas credênciais estão inválidas.",
@@ -7,7 +7,7 @@
     "Você está tentando acessar um app apenas para administradores!": "Você está tentando acessar um app apenas para administradores!",
     "Você está tentando acessar um app apenas para administradores ou produtores!": "Você está tentando acessar um app apenas para administradores ou produtores!",
     "Você está tentando acessar um app apenas para administradores ou consumidores!": "Você está tentando acessar um app apenas para administradores ou consumidores!",
-    "O usuário não está verificado.": "Sua conta não está autenticada. Por favor, verifique o seu e-mail.",
+    "O usuário não está verificado.": "Sua conta não está verifica. Por favor, verifique o seu e-mail.",
     "Não autorizado": "Acesso não autorizado.",
     "O agronegócio não está ativo.": "O seu agronegócio não está ativo! Por favor entre em contato com o CDD.",
     "Essa conta não tem uma senha definida.": "Sua conta não possui uma senha definida.",
